chore(kv): move migrator tests into testing package and run for inmem and bolt
