+++ kv/migration_private_test.go
@@ -1,12 +1,12 @@
 package kv
 
 import (
-	"testing"
 	"time"
 )
 
-func MigratorSetNow(t *testing.T, migrator *Migrator, now func() time.Time) {
-	t.Helper()
-
+// MigratorSetNow sets the now function on the migrator.
+// This function is only reachable via tests defined within this
+// package folder.
+func MigratorSetNow(migrator *Migrator, now func() time.Time) {
 	migrator.now = now
 }
